@@ -59,8 +59,8 @@ public interface PlatformPlugin<C> {
     default ThreadFactory getThreadFactory() {
         return new ThreadFactoryBuilder()
                 .setNameFormat(getName() + " Pool Thread #%1$d")
-                // Hikari create daemons by default and we could daemon threads for our own scheduler too
-                // because we safely shutdown
+                // Hikari create daemons by default. We could daemon threads for our own scheduler too
+                // because we safely shut down
                 .setDaemon(true)
                 .build();
     }
